+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented here. The format is based
 - Fixed a bug that prevented model rollbacks from working correctly
 - Added (rudamentary) ability to plot muscle moment arms while editing a model (related: #191)
 - Updated underlying OpenSim library version v4.3 (#192)
+- Added "OpenSim" muscle coloring option, which uses `GeometryPath::getColor` (#193)
 
 
 ## [0.1.5] - 2022/04/11